*Initial work to update to 4.0.6 Spell, map and other DBC updates and new parsers

+++ Core/WCell.Constants/Misc.cs
@@ -305,7 +305,7 @@ namespace WCell.Constants
 	}
 
 	[Flags]
-	public enum ShapeshiftMask : uint
+	public enum ShapeshiftMask : ulong
 	{
 		Cat = 1,
 		TreeOfLife = 2,
